Create Receive and Transmit
This program implements the simple data transmission and reception process via UART on the Tiva board i.e. without using interrupts.
It also provides visual feedback at various stages of the program execution using the RGB LED on the board. This can be used to debug the program if needed.